[SPARK-33040][FOLLOW-UP][R] Reorder argument choices and add examples
### What changes were proposed in this pull request?

- Reorder choices of `dtype` to match Scala defaults.
- Add example to ml_functions.

### Why are the changes needed?

As requested:

- https://github.com/apache/spark/pull/29917#pullrequestreview-501715344
- https://github.com/apache/spark/pull/29917#pullrequestreview-501716521

### Does this PR introduce _any_ user-facing change?

No (changes to newly added component).

### How was this patch tested?

Existing tests.

Closes #29944 from zero323/SPARK-33040-FOLLOW-UP.

Authored-by: zero323 <mszymkiewicz@gmail.com>

+++ R/pkg/R/functions.R
@@ -354,6 +354,11 @@ NULL
 #' @name column_ml_functions
 #' @rdname column_ml_functions
 #' @family ml functions
+#' @examples
+#' \dontrun{
+#' df <- read.df("data/mllib/sample_libsvm_data.txt", source = "libsvm")
+#' head(select(df, vector_to_array(df$features)))
+#' }
 NULL
 
 #' @details
@@ -4481,7 +4486,7 @@ setMethod("timestamp_seconds",
 #' @note vector_to_array since 3.1.0
 setMethod("vector_to_array",
           signature(x = "Column"),
-          function(x, dtype = c("float32", "float64")) {
+          function(x, dtype = c("float64", "float32")) {
             dtype <- match.arg(dtype)
             jc <- callJStatic(
               "org.apache.spark.ml.functions",
